// frontend/spa/src/app/models/user_manage/user.model.ts
export interface User {
    username: string;
    email: string;
    //nacionality: number;
    //birthdate: Date;
    //weight: number;
    password: string;
    //role: 'Coach' | 'Player';
    //role: string;
    //status: string;
    //created_at: Date;
    //first_name: string;
    //last_name: string;
}